requirements.txt
pyrogram
tgcrypto
requests
python-dotenv

motor
aiofiles
dnspython

google-generativeai
ipython

pillow
ytthumb
googletrans==4.0.0-rc1
countryinfo

pyqrcode
pypng
pyzbar
pyzbar[scripts]